Correcion de todo lo que estaba mal en eliminar

+++ logEliminarSucursal.php
@@ -14,6 +14,7 @@ class Sucursal
         if ($con) {
             //obtener el ID del URL
             $id = $_GET['id'];
+            $deshabilita = false;
             //echo $id. '<br>';
 
             //verificar si la sucursal tiene registros en ventas
@@ -24,15 +25,8 @@ class Sucursal
             $resultados_ventas = sqlsrv_query($con, $querry_ventas);
 
             if (sqlsrv_fetch_array($resultados_ventas)) {
-                echo '<script>alert("Esta sucursal todavía tiene ventas registradas, necesarias para el balance, por lo que no puede ser eliminada. La sucursal será deshabilitada");</script>';
-                $querry_update = "UPDATE sucursal
-                SET Estado = 'Dshabilitado'
-                WHERE id_sucursal = '$id'";
-
-                $stm_des = sqlsrv_query($con, $querry_update);
-
-                echo '<script>alert("La sucursal '. $id . ' ha sido deshabilitada")</script>';
-                include("lista_sucursal.php");
+                //echo '<script>alert("Esta sucursal todavía tiene ventas registradas, necesarias para el balance, por lo que no puede ser eliminada. La sucursal será deshabilitada");</script>';
+                $deshabilita = true;
             } 
             
             else {
@@ -44,51 +38,39 @@ class Sucursal
                 $resultados_inventario = sqlsrv_query($con, $querry_inventario);
 
                 if (sqlsrv_fetch_array($resultados_inventario)) {
-                    echo '<script>alert("Todos los productos que tiene la sucursal '. $id. ' serán eliminados del inventario");</script>';
-                    $querry_delete_inventario = "DELETE FROM Inventario_suc
-                    WHERE id_sucursal = '$id'";
-
-                    $stm_inv = sqlsrv_query($con, $querry_delete_inventario);
+                    $deshabilita = true;
                 }
 
-                //verifica si hay empleados registrados en la sucursal
-                $querry_empleados  = "SELECT id_empleado FROM Empleados
-                    WHERE sucursal = '$id'";
-                //echo '<script>alert('. $id. ');</script>';
-
-                $resultados_empleados = sqlsrv_query($con, $querry_empleados);
-                    $c = 0;
-                    //elimina a los empleados de la tabla de permisoso
-                    while($row = sqlsrv_fetch_array($resultados_empleados)){
-                        $c++;
-                        echo '<script>alert("si entra al ciclo");</script>';
-                        $id_empleado = $row["id_empleado"];
-                        echo '<script>alert("'.$id_empleado.'");</script>';
-
-                        $querry_delete_permisos = "DELETE FROM Permisos
-                            WHERE id_empleado = '$id_empleado'";
-
-                        $stm_per = sqlsrv_query($con, $querry_delete_permisos);
-                    }
-
-                    if($c != 0){
-                        echo '<script>alert("Todos los empleados de la sucursal ' . $id . ' serán eliminados del registro");</script>';
-                    }
-
-                    //elimina a los empleados de la tabla de empleados
-                    $querry_delete_empleados = "DELETE FROM Empleados
+                else{
+                    //verifica si hay trabajadores registrados en esa sucursal
+                    $querry_empleados = "SELECT * FROM Empleados 
                     WHERE sucursal = '$id'";
 
-                    $stm_emp = sqlsrv_query($con, $querry_delete_empleados);
-                
+                    $resultados_empleado = sqlsrv_query($con, $querry_empleados);
 
-                //elimina ahora si la sucursal
+                    if(sqlsrv_fetch_array($resultados_empleado)){
+                        $deshabilita = true;
+                    }
+                }
+            }
+
+            //verifica si se encontro alguna tabla, en caso afirmativo deshabilida la sucursal, de lo contrario, la elimina
+            if($deshabilita == true){
+                $querry_update = "UPDATE sucursal
+                SET Estado = 'Deshabilitado' 
+                WHERE id_sucursal = '$id'";
+
+                $stm = sqlsrv_query($con, $querry_update);
+            }
+
+            else{
                 $querry_delete = "DELETE FROM sucursal
                 WHERE id_sucursal = '$id'";
+
                 $stm = sqlsrv_query($con, $querry_delete);
-                echo '<script>alert("La sucursal y todos los datos relacionados con ella han sido eliminados correctamente");</script>';
-                include_once("lista_sucursal.php");
             }
+
+            include_once("lista_sucursal.php");
 
         }
